pyproject.toml
[project]
name = "netrapier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dearpygui>=2.0.0",
    "nuitka>=2.6.8",
    "psutil>=7.0.0",
    "pyqtdarktheme>=2.1.0",
    "pyside6>=6.8.2.1",
    "python-docx>=1.1.2",
    "requests>=2.32.3",
    "scapy>=2.6.1",
]
